final/requirements.txt
Flask
Flask-SQLAlchemy
flask-restful
flask-security-too
numpy
matplotlib